Fixed SPQuad resizing bug
I discovered this bug when I created an SPTextField with height 0 and then set it to a height > 0. It wasn't visible and it reported its height as 0. I traced this issue to SPQuad. If you create a quad with width or height = 0 and then set it to > 0 later, it does not update. The quad is effectively invisible. One would expect that the quad would be resized correctly even when initialized with height = 0. This change fixes that bug by resetting the vertexData involving width and height when the width or height changes.